fix: file validation pipe
