+++ project/apps/file-vault/src/app/file-uploader/file-uploader.constant.ts
@@ -1,6 +1,5 @@
-export const MB = 1024 * 1024;
-export const ALLOWED_MIMETYPES = ['jpeg', 'jpg', 'png'];
-export const FileMaxSize = {
-  Avatar: 0.5 * MB,
-  PostPhoto: 1 * MB,
-} as const;
+import { Files } from '@project/shared/core';
+
+const { ALLOWED_IMG_MIMETYPES, FileMaxSize } = Files;
+
+export { ALLOWED_IMG_MIMETYPES, FileMaxSize };
