+++ project/apps/file-vault/src/app/file-uploader/file-uploader.controller.ts
@@ -22,7 +22,7 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { FileValidationPipe } from './pipes/file-validation.pipe';
-import { ALLOWED_MIMETYPES, FileMaxSize, MB } from './file-uploader.constant';
+import { ALLOWED_IMG_MIMETYPES, FileMaxSize } from './file-uploader.constant';
 
 @ApiTags('files')
 @Controller('files')
@@ -54,7 +54,9 @@ export class FileUploaderController {
   @Post('/upload/avatar')
   @UseInterceptors(FileInterceptor('avatar'))
   public async uploadAvatarFile(
-    @UploadedFile(new FileValidationPipe(FileMaxSize.Avatar, ALLOWED_MIMETYPES))
+    @UploadedFile(
+      new FileValidationPipe(FileMaxSize.Avatar, ALLOWED_IMG_MIMETYPES)
+    )
     file: Express.Multer.File
   ) {
     const existFile = await this.fileUploaderService.saveFile(file);
@@ -82,7 +84,9 @@ export class FileUploaderController {
   @Post('/upload/post/photo')
   @UseInterceptors(FileInterceptor('photo'))
   public async uploadPostPhotoFile(
-    @UploadedFile(new FileValidationPipe(1 * MB, ALLOWED_MIMETYPES))
+    @UploadedFile(
+      new FileValidationPipe(FileMaxSize.PostPhoto, ALLOWED_IMG_MIMETYPES)
+    )
     file: Express.Multer.File
   ) {
     const existFile = await this.fileUploaderService.saveFile(file);
