+++ project/libs/shared/core/src/index.ts
@@ -3,3 +3,5 @@ export * from './lib/repository/entity.interface';
 export * from './lib/repository/base-memory.repository';
 export * from './lib/repository/base-mongo.repository';
 export * from './lib/pipes/mongo-id-validation.pipe';
+export * from './lib/pipes/file-validation.pipe';
+export * as Files from './lib/consts/files';
