+++ project/libs/shared/core/src/lib/pipes/file-validation.pipe.ts
@@ -1,0 +1,29 @@
+import { Injectable, PipeTransform, BadRequestException } from '@nestjs/common';
+import { extension } from 'mime-types';
+import 'multer';
+
+@Injectable()
+export class FileValidationPipe implements PipeTransform {
+  constructor(
+    private readonly maxSize: number,
+    private readonly allowedMimeTypes: string[]
+  ) {}
+
+  transform(file: Express.Multer.File) {
+    if (!file) {
+      throw new BadRequestException('File is not send.');
+    }
+    const { size, mimetype, originalname } = file;
+    const fileExtension = extension(mimetype) || originalname.split('.').pop();
+
+    if (!fileExtension || !this.allowedMimeTypes.includes(fileExtension)) {
+      throw new BadRequestException('Invalid file format.');
+    }
+
+    if (size > this.maxSize) {
+      throw new BadRequestException('File size exceeds the allowed limit.');
+    }
+
+    return file;
+  }
+}
